Refactor: + ShfApplication factory, seed_helpers is just constants, errors
